feat: convert Match.js to TypeScript
